backend/requirements.txt
bcrypt==4.1.2
email-validator==2.1.0
fastapi==0.129.0
langchain==1.2.10
langchain-google-genai==4.2.0
prisma==0.15.0
pyjwt==2.8.0
python-dotenv==1.0.0
uvicorn[standard]==0.40.0